Report the original stream:error stanza to clients.
When a stream:error was received, the error was parsed, converted to a
StreamError exception, then XML was recreated from that.  This is roundabout
and lossy: it's parsing the XML, then attempting to recreate the same XML
that it already had.

Instead, keep the original XML node in the exception, and use it directly.
This ensures that the <stream:error> node in remote-stream-error is
unchanged from what we received from the remote server.  Any extensions
will be passed through unchanged, as should happen for all stanzas.

For example, the <stream:error> node reported by Jabber14 now correctly
reproduces its xmlns='jabber:client' attribute.

The simple patch in patches.py should eventually be sent upstream.