Fix xep124.py.testWhiteListError.
After the previous commit, this test was succeeding when it shouldn't.
It expected a failure, but didn't check that the error it received was
the correct one; it was receiving the wrong error, which it accepted.
Check that the returned error really is 400 Bad Request.

Also, fix the actual error: there was no route= in this test's request,
so it wasn't routed to the test server, ending up on the default port
5222 instead.