Improve error handling when httpb_client receives a terminate.
On terminate, QueryFactory raised a failure using a raw tuple containing
the parsed stanza.  This leads to obscure error messages that look like
this:

> Traceback (most recent call last):
> Failure: __builtin__.tuple: (<twisted.words.xish.domish.Element object
> at 0x925f72c>, [<twisted.words.xish.domish.Element object at 0x9240cac>])

Stash the data in an exception class, which is what failure.Failure
expects anyway, and update the test that uses it.

(testStreamError seems to be the only place that uses this; I just ended
up cleaning it up in the process of figuring out the above error.)

Now if a <body type='terminate'> is received and printed as a failure, it
looks like this:

> Traceback (most recent call last):
> Failure: punjab.httpb_client.HTTPBNetworkTerminated: <body
> xmlns='http://jabber.org/protocol/httpbind' type='terminate' condition='remote-stream-error'/>