Fix errors in testSessionTimeout cause long timeouts.
Errors in this test are being raised from within callbacks which
are being fired on Deferred objects other than the one returned
from the test itself.  This causes the test runner to never see
them, so the test sits around until the whole test times out.

Fix this by adding an errback to the two separate Deferred objects,
which simply forwards the error onto the primary Deferred.

Lower the timeout from 10*2 seconds to 2*2 seconds.  This is an
improvement, but the test is still cumbersome to run.  Maybe tests
that depend on timeouts should be moved to their own test class,
so it's possible to run the fast tests quickly.