Fixup testSessionTimeout.
The actual problem here was that tests are expected to chain by
returning Deferreds, but the Deferred objects weren't being returned.
The part that tripped me up before was that reactor.callLater doesn't
return a Deferred; task.deferLater does that.  Once that's used, the
callbacks and errors propagate like they're supposed to, normal
assertions can be used, and no weird error propagation hacks are needed.