Fix error handling when invalid XML is received from the server.
An exception was being raised on XML parse errors on
'getattr(streamerror.value, "element")', because streamerror isn't
necessarily a StreamError.  If it's anything else, it's not a
stream:error.  Fix the exception.  Send remote-connection-failed
instead of remote-stream-error; this is a connection failure akin
to losing the TCP connection, not an encapsulated stream:error
from the XMPP server.  Add a test.