Partial fix for legacy port 5223 SSL connections.
xmlstream.ssl exist, so 'if port == 5223 and xmlstream.ssl' was
throwing an exception.  self.connectFunc is actually connectFuncName.
self.connectFuncArgs should be a tuple.

Support "ssl.supported", which is documented in twisted.internet.ssl.
(This is completely braindamaged and ignores all standard conventions
for Python modules...)

I'm not sure that this is completely fixed, as I don't have a port 5223
XMPP server to test with right now, but it's at least attempting to
connect and not failing outright.