Fix server terminations not being reported reliably.
If the XMPP server closes the connection when there's no active
connection to the client, the session was abruptly closed, causing an
"unknown SID" condition the next time the client connects.

Like streamError, if connectError happens while we have no outstanding
requests, store the error and report it when we do.

Call stopTrying explicitly.  This is done when expire() is called, but we
may not call expire().

(The "pending error" logic matches streamError exactly, and they should be
merged.)