Update the twisted plugin to not use mktap.
This uses the mechanism shown here:

http://www.opendocs.net/python/twisted/howto/tap.html

punjab.py had to be renaemd to punjab_plugin.py, because punjab.py
is unable to import the "punjab" module--it just pulls itself in.
Relative imports don't help here, since twisted/plugins isn't
a module.

This also eliminates a deprecation warning when running the server.