Replace .tac instructions.
It's much simpler to run the server by running it directly as a plugin,
instead of messing with a .tac file, which almost nobody really needs to
do.  For people not familiar with Twisted (myself, until a few days ago),
having it in here is confusing, since we don't know the normal way of
running the application.

Maybe the .tac file should be removed.