Add BOSH nack for undelivered messages
XEP-0206 provides for an optional error response when
the BOSH endpoint fails to deliver a stanza to the client.
This patch adds the writing of an appropriate error response
for message and iq stanzas back to the stream if they are not
delivered by the time a session disconnect is triggered.