Add optional strports input to the twistd plugin
The t.a.strports module provides a parser for the simplified string
description of listening ports that is used to generate Endpoints. This
allows services to restart and change the listening connection settings
without needing to adjust the code that returns a Service implementation. It
also standardises the configuration of SSL/TLS listening ports so custom options
are not required to collect keys or interface bindings.